Refactor pickerview
