+++ C7FIT/ProfileViewController.swift
@@ -352,99 +352,3 @@ extension ProfileViewController: UIImagePickerControllerDelegate, UINavigationCo
         dismiss(animated: true, completion: nil)
     }
 }
-
-// MARK: - PickerCellDelegate
-
-/// Adapter Pattern for UIPickerView and UITableViewCells
-extension ProfileViewController: PickerCellDelegate {
-
-    func onPickerOpen(cell: AbstractHealthCell, pickerView: UIPickerView) {
-        switch cell.picker.tag {
-        // Start first two cases at weight 150 lbs and height 6'5" for faster navigation
-        case 1:
-            if cell.dataLabel.text!.isEmpty {
-                pickerView.selectRow(150, inComponent: 0, animated: true)
-                cell.dataLabel.text = ProfileViewModel.personalWeight[150]
-                updateBMI()
-            } else {
-                cell.dataLabel.text = cell.dataLabel.text
-            }
-        case 2:
-            if cell.dataLabel.text!.isEmpty {
-                pickerView.selectRow(54, inComponent: 0, animated: true)
-                cell.dataLabel.text = ProfileViewModel.personalHeight[54]
-                updateBMI()
-            } else {
-                cell.dataLabel.text = cell.dataLabel.text
-            }
-        case 5, 6:
-            cell.dataLabel.text = cell.dataLabel.text!.isEmpty ? ProfileViewModel.repetitions[0] : cell.dataLabel.text
-        case 7, 8, 9:
-            cell.dataLabel.text = cell.dataLabel.text!.isEmpty ? ProfileViewModel.weights[0] : cell.dataLabel.text
-        default:
-            break
-        }
-        cell.dataLabel.textColor = .red
-    }
-
-    func onPickerClose(_ cell: AbstractHealthCell) {
-        cell.dataLabel.textColor = .black
-    }
-
-    func pickerView(_ pickerView: UIPickerView, titleForRow row: Int,
-                    forComponent component: Int, forCell cell: AbstractHealthCell) -> String? {
-        switch pickerView.tag {
-        case 1:
-            return ProfileViewModel.personalWeight[row]
-        case 2:
-            return ProfileViewModel.personalHeight[row]
-        case 5, 6:
-            return ProfileViewModel.repetitions[row]
-        case 7, 8, 9:
-            return ProfileViewModel.weights[row]
-        default:
-            return nil
-        }
-    }
-
-    func pickerView(_ pickerView: UIPickerView, didSelectRow row: Int,
-                    inComponent component: Int, forCell cell: AbstractHealthCell) {
-        switch pickerView.tag {
-        case 1:
-            cell.dataLabel.text = ProfileViewModel.personalWeight[row]
-            updateBMI()
-        case 2:
-            cell.dataLabel.text = ProfileViewModel.personalHeight[row]
-            updateBMI()
-        case 5, 6:
-            cell.dataLabel.text = ProfileViewModel.repetitions[row]
-        case 7, 8, 9:
-            cell.dataLabel.text = ProfileViewModel.weights[row]
-        default:
-            break
-        }
-    }
-}
-
-// MARK: - PickerCellDataSource
-
-extension ProfileViewController: PickerCellDataSource {
-    public func numberOfComponents(in pickerView: UIPickerView, forCell cell: AbstractHealthCell) -> Int {
-        return 1
-    }
-
-    func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int, forCell: AbstractHealthCell) -> Int {
-        switch pickerView.tag {
-        case 1:
-            return ProfileViewModel.personalWeight.count
-        case 2:
-            return ProfileViewModel.personalHeight.count
-        case 5, 6:
-            return ProfileViewModel.repetitions.count
-        case 7, 8, 9:
-            return ProfileViewModel.weights.count
-        default:
-            return 0
-        }
-    }
-}
